+++ opennmt/tests/reducer_test.py
@@ -7,6 +7,42 @@ from opennmt.layers import reducer
 
 class ReducerTest(tf.test.TestCase):
 
+  def testAlignInTimeSame(self):
+    a = [
+        [[1], [0], [0]],
+        [[1], [2], [3]]]
+    length = 3
+    b = reducer.align_in_time(tf.constant(a, dtype=tf.float32), tf.constant(length))
+    self.assertEqual(1, b.get_shape().as_list()[-1])
+    with self.test_session() as sess:
+      self.assertAllEqual(a, sess.run(b))
+
+  def testAlignInTimeLarger(self):
+    a = [
+        [[1], [0], [0]],
+        [[1], [2], [3]]]
+    length = 4
+    b = [
+        [[1], [0], [0], [0]],
+        [[1], [2], [3], [0]]]
+    c = reducer.align_in_time(tf.constant(a, dtype=tf.float32), tf.constant(length))
+    self.assertEqual(1, c.get_shape().as_list()[-1])
+    with self.test_session() as sess:
+      self.assertAllEqual(b, sess.run(c))
+
+  def testAlignInTimeSmaller(self):
+    a = [
+        [[1], [0], [0]],
+        [[1], [2], [0]]]
+    length = 2
+    b = [
+        [[1], [0]],
+        [[1], [2]]]
+    c = reducer.align_in_time(tf.constant(a, dtype=tf.float32), tf.constant(length))
+    self.assertEqual(1, c.get_shape().as_list()[-1])
+    with self.test_session() as sess:
+      self.assertAllEqual(b, sess.run(c))
+
   def testPadWithIdentity(self):
     tensor = [
         [[1], [-1], [-1]],
@@ -31,6 +67,32 @@ class ReducerTest(tf.test.TestCase):
       padded = sess.run(padded)
       self.assertAllEqual(expected, padded)
 
+  def testPadWithIdentityWithMaxTime(self):
+    tensor = [
+        [[1], [-1], [-1], [-1]],
+        [[1], [2], [3], [-1]],
+        [[1], [2], [-1], [-1]]]
+    expected = [
+        [[1], [1], [1], [1], [0], [0]],
+        [[1], [2], [3], [1], [1], [0]],
+        [[1], [2], [0], [0], [0], [0]]]
+    lengths = [1, 3, 2]
+    max_lengths = [4, 5, 2]
+    maxlen = 6
+
+    padded = reducer.pad_with_identity(
+        tf.constant(tensor, dtype=tf.float32),
+        tf.constant(lengths),
+        tf.constant(max_lengths),
+        identity_values=1,
+        maxlen=maxlen)
+
+    self.assertEqual(1, padded.get_shape().as_list()[-1])
+
+    with self.test_session() as sess:
+      padded = sess.run(padded)
+      self.assertAllEqual(expected, padded)
+
   def testPadNWithIdentity(self):
     a = [
         [[1], [-1], [-1]],
@@ -48,6 +110,37 @@ class ReducerTest(tf.test.TestCase):
         [[1], [2], [3], [4]],
         [[1], [2], [1], [0]],
         [[1], [2], [0], [0]]]
+    length_a = [1, 3, 2]
+    length_b = [4, 2, 2]
+
+    (padded_a, padded_b), length = reducer.pad_n_with_identity(
+        [tf.constant(a, dtype=tf.float32), tf.constant(b, dtype=tf.float32)],
+        [tf.constant(length_a), tf.constant(length_b)],
+        identity_values=1)
+
+    with self.test_session() as sess:
+      padded_a, padded_b, length = sess.run([padded_a, padded_b, length])
+      self.assertAllEqual([4, 3, 2], length)
+      self.assertAllEqual(expected_a, padded_a)
+      self.assertAllEqual(expected_b, padded_b)
+
+  def testPadNWithIdentityWithMaxTime(self):
+    a = [
+        [[1], [-1], [-1]],
+        [[1], [2], [3]],
+        [[1], [2], [-1]]]
+    b = [
+        [[1], [2], [3], [4], [-1]],
+        [[1], [2], [-1], [-1], [-1]],
+        [[1], [2], [-1], [-1], [-1]]]
+    expected_a = [
+        [[1], [1], [1], [1], [0]],
+        [[1], [2], [3], [0], [0]],
+        [[1], [2], [0], [0], [0]]]
+    expected_b = [
+        [[1], [2], [3], [4], [0]],
+        [[1], [2], [1], [0], [0]],
+        [[1], [2], [0], [0], [0]]]
     length_a = [1, 3, 2]
     length_b = [4, 2, 2]
 
@@ -103,6 +196,31 @@ class ReducerTest(tf.test.TestCase):
       self.assertAllEqual(expected, reduced)
       self.assertAllEqual([4, 3, 2], length)
 
+  def testMultiplyReducerWithSequenceAndMaxTime(self):
+    a = [
+        [[1], [-1], [-1]],
+        [[1], [2], [3]],
+        [[1], [2], [-1]]]
+    b = [
+        [[1], [2], [3], [4], [-1]],
+        [[1], [2], [-1], [-1], [-1]],
+        [[1], [2], [-1], [-1], [-1]]]
+    expected = [
+        [[1], [2], [3], [4], [0]],
+        [[1], [4], [3], [0], [0]],
+        [[1], [4], [0], [0], [0]]]
+    length_a = [1, 3, 2]
+    length_b = [4, 2, 2]
+
+    reduced, length = reducer.MultiplyReducer().reduce_sequence(
+        [tf.constant(a, dtype=tf.float32), tf.constant(b, dtype=tf.float32)],
+        [tf.constant(length_a), tf.constant(length_b)])
+
+    with self.test_session() as sess:
+      reduced, length = sess.run([reduced, length])
+      self.assertAllEqual(expected, reduced)
+      self.assertAllEqual([4, 3, 2], length)
+
   def testConcatInDepthWithSequence(self):
     a = [
         [[1], [-1], [-1]],
@@ -139,6 +257,33 @@ class ReducerTest(tf.test.TestCase):
         [[1], [2], [3], [4]],
         [[1], [2], [-1], [-1]],
         [[1], [2], [-1], [-1]]]
+    expected = [
+        [[1], [1], [2], [3], [4]],
+        [[1], [2], [3], [1], [2]],
+        [[1], [2], [1], [2], [0]]]
+    length_a = [1, 3, 2]
+    length_b = [4, 2, 2]
+
+    reduced, length = reducer.ConcatReducer(axis=1).reduce_sequence(
+        [tf.constant(a, dtype=tf.float32), tf.constant(b, dtype=tf.float32)],
+        [tf.constant(length_a), tf.constant(length_b)])
+
+    self.assertEqual(1, reduced.get_shape().as_list()[-1])
+
+    with self.test_session() as sess:
+      reduced, length = sess.run([reduced, length])
+      self.assertAllEqual(expected, reduced)
+      self.assertAllEqual([5, 5, 4], length)
+
+  def testConcatInTimeWithSequenceAndMaxTimeMismatch(self):
+    a = [
+        [[1], [-1], [-1]],
+        [[1], [2], [3]],
+        [[1], [2], [-1]]]
+    b = [
+        [[1], [2], [3], [4], [-1], [-1]],
+        [[1], [2], [-1], [-1], [-1], [-1]],
+        [[1], [2], [-1], [-1], [-1], [-1]]]
     expected = [
         [[1], [1], [2], [3], [4]],
         [[1], [2], [3], [1], [2]],